temp fix ui unavailable

+++ features/transaction/cart/src/main/java/com/tokopedia/cartrevamp/data/model/response/shopgroupsimplified/UnavailableSection.kt
@@ -14,5 +14,5 @@ data class UnavailableSection(
     @SerializedName("selected_unavailable_action_id")
     val selectedUnavailableActionId: String = "",
     @SerializedName("products_count")
-    val productsCount: Int = 0
+    val productsCount: Long = 0
 )
